Add resource authorization to delete feature

+++ TutorWebApi.Application/Services/ProfileService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Logging;
 using TutorWebApi.Domain;
 
@@ -10,14 +11,16 @@ namespace TutorWebApi.Application
         private readonly IUserContextService _userContextService;
         private readonly IMapper _mapper;
         private readonly ILogger<ProfileService> _logger;
+        private readonly IAuthorizationService _authorizationService;
 
         public ProfileService(IProfileRepository profileRepository, IUserContextService userContextService
-            , IMapper mapper, ILogger<ProfileService> logger)
+            , IMapper mapper, ILogger<ProfileService> logger, IAuthorizationService authorizationService)
         {
             _profileRepository = profileRepository;
             _userContextService = userContextService;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
         }
 
         public async Task<int> CreateProfile(ProfileDto profilDto)
@@ -25,7 +28,7 @@ namespace TutorWebApi.Application
             var userId = (int)_userContextService.GetUserId();
             var result = _profileRepository.IsUserHaveProfile(userId).Result;
             if (result)
-                throw new BadRequestForCreateException("It is forbidden to create a second profile");
+                throw new ForbidException("It is forbidden to create a second profile");
 
             var profile = _mapper.Map<Domain.Profile>(profilDto);
             profile.UserRef = userId;
@@ -34,16 +37,25 @@ namespace TutorWebApi.Application
             return profileId;
         }
 
-        public async Task DeleteProfile()
+        public async Task DeleteProfile(int profileId)
         {
+            var user = _userContextService.GetUser();
             var userId = (int)_userContextService.GetUserId();
-            _logger.LogInformation($"Profile with userRef:{userId} Delete action invoked");
+            _logger.LogInformation($"Profile id: {profileId} with userRef:{userId} Delete action invoked");
 
             var result = await _profileRepository.IsUserHaveProfile(userId);
             if (!result)
                 throw new NotFoundException("User does not have Profile");
 
-            var profileId = await _profileRepository.GetProfilIdByUser(userId);
+            var profile = await _profileRepository.GetProfileById(profileId);
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(user, profile,
+                new ResourceOperationRequirement(ResourceOperation.Delete));
+
+            if(!authorizationResult.Succeeded)
+            {
+                throw new ForbidException("");
+            }
 
             await _profileRepository.DeleteAllAchievementsByProfile(profileId);
             await _profileRepository.DeleteAllExperiencesByProfile(profileId);
